requirements.txt
opencv-python
google-cloud-storage
psycopg2
